# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cart": {
      "get": {
        "operationId": "CartController_getCart",
        "parameters": [
          {
            "name": "x-cart-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "in": "header",
            "description": "Guest cart id (UUID) generated and stored by the client. Required for cart requests without an authenticated session, and used to merge the guest cart into the user cart on login.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The cart for the current user, or the empty cart for a new guest",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/GetCartResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Malformed X-Cart-Token header",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Get the current cart (user cart or guest cart)",
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart/items": {
      "post": {
        "operationId": "CartController_addCartItem",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "CSRF token from GET /auth/csrf — copy the csrfToken field into this header.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "in": "header",
            "description": "Guest cart id (UUID) generated and stored by the client. Required for cart requests without an authenticated session, and used to merge the guest cart into the user cart on login.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The cart after adding the variant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/AddCartItemResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Out of stock, quantity above stock, or missing guest cart token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Variant not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Add a variant to the cart (merges the guest cart first when logged in)",
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart/items/{variantId}": {
      "patch": {
        "operationId": "CartController_updateCartItem",
        "parameters": [
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "description": "Product variant id",
            "schema": {
              "example": 3744,
              "type": "number"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "CSRF token from GET /auth/csrf — copy the csrfToken field into this header.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "in": "header",
            "description": "Guest cart id (UUID) generated and stored by the client. Required for cart requests without an authenticated session, and used to merge the guest cart into the user cart on login.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartItemRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The cart after updating the quantity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/UpdateCartItemResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Quantity above stock or missing guest cart token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Variant not found or not present in the cart",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Set the quantity of a variant in the cart",
        "tags": [
          "Cart"
        ]
      },
      "delete": {
        "operationId": "CartController_removeCartItem",
        "parameters": [
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "description": "Product variant id",
            "schema": {
              "example": 3744,
              "type": "number"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "CSRF token from GET /auth/csrf — copy the csrfToken field into this header.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Cart-Token",
            "in": "header",
            "description": "Guest cart id (UUID) generated and stored by the client. Required for cart requests without an authenticated session, and used to merge the guest cart into the user cart on login.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The cart after removing the variant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/RemoveCartItemResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing guest cart token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Variant not present in the cart",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Remove a variant from the cart",
        "tags": [
          "Cart"
        ]
      }
    }
  },
  "info": {
    "title": "Lio API",
    "description": "Authentication and session management API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "GetCartProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 3342
          },
          "name": {
            "type": "string",
            "example": "کمل کامپکت آبی ایرانی"
          },
          "slug": {
            "type": "string",
            "example": "kamel-compact-abi"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "GetCartVariantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 3744
          },
          "sku": {
            "type": "string",
            "example": "SEED-SKU-1-1"
          },
          "price": {
            "type": "number",
            "example": 1694000,
            "description": "Current selling price"
          },
          "compareAtPrice": {
            "type": "number",
            "example": 1744000,
            "nullable": true,
            "description": "Reference price before discount"
          },
          "stock": {
            "type": "number",
            "example": 12
          }
        },
        "required": [
          "id",
          "sku",
          "price",
          "compareAtPrice",
          "stock"
        ]
      },
      "GetCartItemDto": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "number",
            "example": 3744,
            "description": "Product variant id. Also used as the :variantId path parameter."
          },
          "quantity": {
            "type": "number",
            "example": 2
          },
          "lineTotal": {
            "type": "number",
            "example": 3388000,
            "description": "Current variant price multiplied by quantity"
          },
          "product": {
            "$ref": "#/components/schemas/GetCartProductDto"
          },
          "variant": {
            "$ref": "#/components/schemas/GetCartVariantDto"
          }
        },
        "required": [
          "variantId",
          "quantity",
          "lineTotal",
          "product",
          "variant"
        ]
      },
      "GetCartResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetCartItemDto"
            }
          },
          "itemCount": {
            "type": "number",
            "example": 3,
            "description": "Sum of all item quantities"
          },
          "distinctItemCount": {
            "type": "number",
            "example": 2,
            "description": "Number of distinct variants in the cart"
          },
          "subtotal": {
            "type": "number",
            "example": 3388000,
            "description": "Sum of the line totals, using current variant prices"
          }
        },
        "required": [
          "items",
          "itemCount",
          "distinctItemCount",
          "subtotal"
        ]
      },
      "AddCartItemRequestDto": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "number",
            "example": 3744,
            "description": "Product variant id to add"
          },
          "quantity": {
            "type": "number",
            "example": 1,
            "default": 1,
            "minimum": 1,
            "description": "How many units to add. Defaults to 1."
          }
        },
        "required": [
          "variantId"
        ]
      },
      "AddCartItemProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 3342
          },
          "name": {
            "type": "string",
            "example": "کمل کامپکت آبی ایرانی"
          },
          "slug": {
            "type": "string",
            "example": "kamel-compact-abi"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "AddCartItemVariantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 3744
          },
          "sku": {
            "type": "string",
            "example": "SEED-SKU-1-1"
          },
          "price": {
            "type": "number",
            "example": 1694000,
            "description": "Current selling price"
          },
          "compareAtPrice": {
            "type": "number",
            "example": 1744000,
            "nullable": true,
            "description": "Reference price before discount"
          },
          "stock": {
            "type": "number",
            "example": 12
          }
        },
        "required": [
          "id",
          "sku",
          "price",
          "compareAtPrice",
          "stock"
        ]
      },
      "AddCartItemItemDto": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "number",
            "example": 3744,
            "description": "Product variant id. Also used as the :variantId path parameter."
          },
          "quantity": {
            "type": "number",
            "example": 2
          },
          "lineTotal": {
            "type": "number",
            "example": 3388000,
            "description": "Current variant price multiplied by quantity"
          },
          "product": {
            "$ref": "#/components/schemas/AddCartItemProductDto"
          },
          "variant": {
            "$ref": "#/components/schemas/AddCartItemVariantDto"
          }
        },
        "required": [
          "variantId",
          "quantity",
          "lineTotal",
          "product",
          "variant"
        ]
      },
      "AddCartItemResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddCartItemItemDto"
            }
          },
          "itemCount": {
            "type": "number",
            "example": 3,
            "description": "Sum of all item quantities"
          },
          "distinctItemCount": {
            "type": "number",
            "example": 2,
            "description": "Number of distinct variants in the cart"
          },
          "subtotal": {
            "type": "number",
            "example": 3388000,
            "description": "Sum of the line totals, using current variant prices"
          }
        },
        "required": [
          "items",
          "itemCount",
          "distinctItemCount",
          "subtotal"
        ]
      },
      "UpdateCartItemRequestDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "example": 2,
            "minimum": 1,
            "description": "New absolute quantity for the variant."
          }
        },
        "required": [
          "quantity"
        ]
      },
      "UpdateCartItemProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 3342
          },
          "name": {
            "type": "string",
            "example": "کمل کامپکت آبی ایرانی"
          },
          "slug": {
            "type": "string",
            "example": "kamel-compact-abi"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "UpdateCartItemVariantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 3744
          },
          "sku": {
            "type": "string",
            "example": "SEED-SKU-1-1"
          },
          "price": {
            "type": "number",
            "example": 1694000,
            "description": "Current selling price"
          },
          "compareAtPrice": {
            "type": "number",
            "example": 1744000,
            "nullable": true,
            "description": "Reference price before discount"
          },
          "stock": {
            "type": "number",
            "example": 12
          }
        },
        "required": [
          "id",
          "sku",
          "price",
          "compareAtPrice",
          "stock"
        ]
      },
      "UpdateCartItemItemDto": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "number",
            "example": 3744,
            "description": "Product variant id. Also used as the :variantId path parameter."
          },
          "quantity": {
            "type": "number",
            "example": 2
          },
          "lineTotal": {
            "type": "number",
            "example": 3388000,
            "description": "Current variant price multiplied by quantity"
          },
          "product": {
            "$ref": "#/components/schemas/UpdateCartItemProductDto"
          },
          "variant": {
            "$ref": "#/components/schemas/UpdateCartItemVariantDto"
          }
        },
        "required": [
          "variantId",
          "quantity",
          "lineTotal",
          "product",
          "variant"
        ]
      },
      "UpdateCartItemResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateCartItemItemDto"
            }
          },
          "itemCount": {
            "type": "number",
            "example": 3,
            "description": "Sum of all item quantities"
          },
          "distinctItemCount": {
            "type": "number",
            "example": 2,
            "description": "Number of distinct variants in the cart"
          },
          "subtotal": {
            "type": "number",
            "example": 3388000,
            "description": "Sum of the line totals, using current variant prices"
          }
        },
        "required": [
          "items",
          "itemCount",
          "distinctItemCount",
          "subtotal"
        ]
      },
      "RemoveCartItemProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 3342
          },
          "name": {
            "type": "string",
            "example": "کمل کامپکت آبی ایرانی"
          },
          "slug": {
            "type": "string",
            "example": "kamel-compact-abi"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
...
      "RemoveCartItemItemDto": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "number",
            "example": 3744,
            "description": "Product variant id. Also used as the :variantId path parameter."
          },
          "quantity": {
            "type": "number",
            "example": 2
          },
          "lineTotal": {
            "type": "number",
            "example": 3388000,
            "description": "Current variant price multiplied by quantity"
          },
          "product": {
            "$ref": "#/components/schemas/RemoveCartItemProductDto"
          },
          "variant": {
            "$ref": "#/components/schemas/RemoveCartItemVariantDto"
          }
        },
        "required": [
          "variantId",
          "quantity",
          "lineTotal",
          "product",
          "variant"
        ]
      },
      "RemoveCartItemResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RemoveCartItemItemDto"
            }
          },
          "itemCount": {
            "type": "number",
            "example": 3,
            "description": "Sum of all item quantities"
          },
          "distinctItemCount": {
            "type": "number",
            "example": 2,
            "description": "Number of distinct variants in the cart"
          },
          "subtotal": {
            "type": "number",
            "example": 3388000,
            "description": "Sum of the line totals, using current variant prices"
          }
        },
        "required": [
          "items",
          "itemCount",
          "distinctItemCount",
          "subtotal"
        ]
      }
    }
  }
}